Fix missed renames
